+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 - **[Breaking change]** Replace `OutModules` enum by custom compiler option `mjsModule`.
 - **[Breaking change]** Drop support for Pug, Sass, Angular & Webpack.
 - **[Feature]** Expose custom registries for each target.
+- **[Feature]** Add `dist.tscOptions` for `lib` target to override options for
+  distribution builds.
 - **[Fix]** Disable deprecated TsLint rules from the default config
 - **[Fix]** Remove use of experimental `fs/promises` module.
 - **[Internal]** Fix continuous deployment script (stop confusing PRs to master